requirements.txt
openai>=1.14.0
google-api-python-client>=2.120.0
google-auth>=2.28.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.6.3
pydantic-settings>=2.2.1
loguru>=0.7.2
tenacity>=8.2.3
tinydb>=4.8.0